# openapi.yaml
openapi: 3.0.3
info:
  title: Corgi Recommender Service API
  description: |
    API for the Corgi Recommender Service, a personalized recommendations system
    for Mastodon posts. The API handles user interactions, preferences, and generates
    personalized timelines.
    
    ## Key Features
    - **User Interaction Tracking**: Records favorites, bookmarks, and explicit feedback
    - **Personalized Recommendations**: Generates tailored post suggestions based on user activity
    - **Privacy Controls**: Provides user-configurable privacy settings
    - **Mastodon Compatibility**: Returns data in Mastodon-compatible format for easy integration
    - **Transparent Proxy**: Proxies requests to Mastodon instances with added recommendations
  version: 1.0.0
  contact:
    name: Corgi Service Team
    email: support@corgi-recommender.example.com
servers:
  - url: http://localhost:5001
    description: Local development server
  - url: https://api.corgi-recommender.example.com
    description: Production server
...
paths:
...
  /api/v1/timelines/home:
    get:
      summary: Get user's home timeline
      description: |
        Returns the user's home timeline, similar to a Mastodon home timeline.
        For real users, the request is proxied to their Mastodon instance.
        For test/synthetic users, returns mock data.
      operationId: getHomeTimeline
      tags:
        - timelines
        - proxy
      parameters:
        - name: user_id
          in: query
          required: false
          schema:
            type: string
          description: The user ID to get timeline for (for testing/validation)
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
          description: Maximum number of posts to return
        - name: strategy
          in: query
          required: false
          schema:
            type: string
            enum: [uniform, after_n, first_only, tag_match]
          description: Injection strategy to use for recommended posts
        - name: inject
          in: query
          required: false
          schema:
            type: boolean
            default: true
          description: Whether to inject recommended posts into the timeline
      responses:
        '200':
          description: Home timeline posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MastodonPost'
                description: Array of Mastodon-compatible posts, possibly with injected recommendations
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  schemas:
...
    MastodonPost:
      type: object
      required:
        - id
        - content
        - created_at
        - account
      properties:
        id:
          type: string
          description: Unique identifier for the post
        content:
          type: string
          description: HTML content of the post
        created_at:
          type: string
          format: date-time
          description: Creation timestamp in ISO 8601 format
        language:
          type: string
          description: ISO language code (e.g., "en")
        account:
          type: object
          required:
            - id
            - username
            - display_name
          properties:
            id:
              type: string
              description: Unique identifier for the account
            username:
              type: string
              description: Username of the account
            display_name:
              type: string
              description: Display name of the account
            followers_count:
              type: integer
              description: Number of followers
            following_count:
              type: integer
              description: Number of accounts following
            statuses_count:
              type: integer
              description: Number of posted statuses
            url:
              type: string
              format: uri
              description: URL to the account's profile
        replies_count:
          type: integer
          description: Number of replies to this post
        reblogs_count:
          type: integer
          description: Number of reblogs of this post
        favourites_count:
          type: integer
          description: Number of favorites for this post
        url:
          type: string
          format: uri
          description: URL to the post
        sensitive:
          type: boolean
          description: Whether the post contains sensitive content
        spoiler_text:
          type: string
          description: Content warning text
        visibility:
          type: string
          enum: [public, unlisted, private, direct]
          description: Visibility level of the post
        ranking_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Recommendation ranking score (0.0 to 1.0)
        recommendation_reason:
          type: string
          description: Human-readable reason for recommendation
        is_recommendation:
          type: boolean
          description: Whether this post was added as a recommendation
        is_real_mastodon_post:
          type: boolean
          description: Whether this is a real Mastodon post (vs. synthetic)
        is_synthetic:
          type: boolean
          description: Whether this is a synthetic post
        injected:
          type: boolean
          description: Whether this post was injected by the recommendation system
        injection_metadata:
          type: object
          description: Metadata about the injected recommendation post
          properties:
            source:
              type: string
              description: Source of the recommendation (e.g., recommendation_engine, timeline_injector)
              example: recommendation_engine
            strategy:
              type: string
              description: Strategy used for the recommendation (e.g., personalized, tag_match)
              example: personalized
            explanation:
              type: string
              description: Human-readable explanation for why this post was recommended
              example: Suggested based on your recent interests in #python and #opensource
            score:
              type: number
              format: float
              description: Recommendation confidence score (0-1)
              example: 0.87
...
  responses:
    BadRequest:
      description: Bad request due to invalid parameters or payload
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              received:
                type: object
                additionalProperties: true
...
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              type:
                type: string
              request_id:
                type: string
...